+++ src/core/base/model/entity.model.ts
