+++ src/core/base/repository/repository.interface.ts
