+++ src/core/exercise/model/exercise.model.ts
