--- src/core/exercise/repositories/exercise.repository.interface.ts
@@ -1,4 +1,0 @@
-import { RepositoryInterface } from 'src/core/base/repository.interface'
-import Exercise from 'src/core/exercise/models/exercise.model'
-
-export type ExerciseRepositoryInterface = RepositoryInterface<Exercise>
