+++ src/core/exercise/repository/exercise-mock.repository.ts
@@ -1,6 +1,6 @@
-import { deleteActionResponse, FindParams } from 'src/core/base/repository.interface'
-import Exercise from 'src/core/exercise/models/exercise.model'
-import ExerciseRepositoryInterface from 'src/core/exercise/repositories/exercise.repository.interface'
+import { deleteActionResponse, FindParams } from 'src/core/base/repository/repository.interface'
+import Exercise from 'src/core/exercise/model/exercise.model'
+import ExerciseRepositoryInterface from 'src/core/exercise/repository/exercise.repository.interface'
 import { exercisesMock } from 'src/mocks/exercises'
 
 export default class ExerciseMockRepository implements ExerciseRepositoryInterface {
