+++ src/core/exercise/repository/exercise.repository.interface.ts
@@ -1,0 +1,4 @@
+import { RepositoryInterface } from 'src/core/base/repository/repository.interface'
+import Exercise from 'src/core/exercise/model/exercise.model'
+
+export type ExerciseRepositoryInterface = RepositoryInterface<Exercise>
