+++ src/core/weight/model/weight.model.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest'
 import { createWeight } from 'src/core/weight/dtos/create-weight.dto'
 import { faker } from '@faker-js/faker'
-import { WeightModel } from 'src/core/weight/models/weight.model'
+import { WeightModel } from 'src/core/weight/model/weight.model'
 describe('Weight Model Test', () => {
   it('Should return an instance of WeightModel', async () => {
     const dataToCreate: createWeight = {
