+++ src/core/weight/model/weight.model.ts
@@ -1,6 +1,6 @@
 import { IsDate, IsNumber } from 'class-validator'
 import { createWeight } from 'src/core/weight/dtos/create-weight.dto'
-import { EntityModelBase } from 'src/core/base/entity.model'
+import { EntityModelBase } from 'src/core/base/model/entity.model'
 
 /**
  * Model representing a weight entry.
