--- src/core/weight/repositories/implementations/weight.firebase.repository.ts
@@ -1,8 +1,0 @@
-import { FirebaseRepository } from 'src/core/base/firebase.repository'
-import { WeightModel } from 'src/core/weight/models/weight.model'
-
-export class WeightFirebaseRepository extends FirebaseRepository<WeightModel> {
-  constructor () {
-    super('weight')
-  }
-}
