--- src/core/weight/repositories/index.ts
@@ -1,3 +1,0 @@
-import { WeightFirebaseRepository } from 'src/core/weight/repositories/implementations/weight.firebase.repository'
-
-export const weightRepositoryInstance = new WeightFirebaseRepository()
