--- src/core/weight/repositories/weight.repository.interface.ts
@@ -1,5 +1,0 @@
-import { RepositoryInterface } from 'src/core/base/repository.interface'
-import { WeightModel } from 'src/core/weight/models/weight.model'
-
-type WeightRepositoryInterface = RepositoryInterface<WeightModel>
-export default WeightRepositoryInterface
