+++ src/core/weight/repository/index.ts
@@ -1,0 +1,3 @@
+import { WeightFirebaseRepository } from 'src/core/weight/repository/implementations/weight.firebase.repository'
+
+export const weightRepositoryInstance = new WeightFirebaseRepository()
