+++ src/core/weight/repository/weight.repository.interface.ts
@@ -1,0 +1,5 @@
+import { RepositoryInterface } from 'src/core/base/repository/repository.interface'
+import { WeightModel } from 'src/core/weight/model/weight.model'
+
+type WeightRepositoryInterface = RepositoryInterface<WeightModel>
+export default WeightRepositoryInterface
